happy ball chat msg

+++ app/src/main/java/com/whzl/mengbi/chat/room/message/messageJson/PrizePoolFullJson.java
@@ -19,7 +19,7 @@ public class PrizePoolFullJson {
          * prizePoolNumber : 380000
          */
 
-        public int countDownSecond;
+        public long countDownSecond;
         public String periodNumber;
         public int prizePoolNumber;
     }
